// src/database/types/case.ts
export interface Case {
  caseName: string;
  number: string;
  displayText: string;
  totalCost: string;
  caseStatus: string;
  caseStatusColor: string;
  contentItemId: string;
  statusId: string;
  parcelNumber: string;
  isManualAddress: string;
  billingStatus: string;
  billingStatusId: string;
  email: string;
  phoneNumber: string;
  caseTypeId: string;
  caseType: string;
  location: string;
  subTypes: string;
  description: string;
  longitudeField: string;
  latitudeField: string;
  caseNumberDetail: string;
  expectedCaseDate: string;
  actualCaseDate: string;
  isEditable: boolean;
  isEdited: boolean;
  isSync: boolean;
  isSubScreenEdited: boolean;
  isForceSync: boolean;
  createdUtc: string;
  modifiedUtc: string;
  DataType: string;
  published: boolean;
  isAllowEditActualDate: boolean;
  viewOnlyAssignUsers: boolean;
  assignedUsers: string;
  author: string;
  ownerName: string;
  IsEnableMultiline: boolean;
  cityField: string;
  countryField: string;
  postalCodeField: string;
  stateField: string;
  streetRouteField: string;
  mailingAddress: string;
  mailingAddressCityField: string;
  mailingAddressCountryField: string;
  mailingAddressPostalCodeField: string;
  mailingAddressStateField: string;
  mailingAddressStreetRouteField: string;
  apartmentSuite: string;
  isShowInspectionTab: boolean;
  caseTag: string;
  isAllowEditCase: boolean;
  isAllowViewInspection: boolean;
  isPermission: boolean;
  isAllowAddAdminNotes: boolean;
  notInOffline: boolean;
  correlationId: string;
  ApiChangeDateUtc: string;
  quickRefNumber: string;
  isStatusReadOnly: boolean;
  isForceSyncSuccess: boolean;
}

export interface SyncQueueItem {
  id: number;
  tableName: string;
  operation: 'INSERT' | 'UPDATE' | 'DELETE';
  data: string;
  createdAt: string;
  status: 'PENDING' | 'SYNCED' | 'FAILED';
}